requirements.txt
absl-py==2.2.2
albumentations==1.3.1
appdirs==1.4.4
asttokens==3.0.0
attrs==25.3.0
autocommand==2.2.2
backcall==0.2.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.6
docopt==0.6.2
efficientnet_pytorch==0.7.1
einops==0.8.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
grad-cam==1.4.8
grpcio==1.71.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.0.0
importlib_resources==6.5.2
inflect==7.3.1
ipython==8.12.3
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mistune==3.1.3
more-itertools==10.3.0
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
onnx==1.17.0
onnxruntime==1.19.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.2.2
prettytable==3.16.0
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.8
Pygments==2.19.1
pyparsing==3.2.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pytorch-wavelets==1.3.0
pytz==2025.1
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq==26.4.0
qudida==0.0.4
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tidecv==1.0.1
tifffile==2024.8.30
timm==0.9.8
tinycss2==1.4.0
tomli==2.0.1
# torch==2.2.2
# torchaudio==2.2.2
# torchvision==0.17.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==2.2.0
ttach==0.0.3
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
yarg==0.1.9
zipp==3.21.0
